+++ frontend/src/pages/CreatedMentoring/components/MentoringApplicationItem/MentoringApplicationItem.tsx
@@ -1,5 +1,6 @@
 import styled from '@emotion/styled';
 
+import ActionButtons from '../ActionButtons/ActionButtons';
 import MentoringApplicationStatus from '../MentoringApplicationStatus/MentoringApplicationStatus';
 import PhoneNumber from '../PhoneNumber/PhoneNumber';
 
@@ -32,12 +33,7 @@ function MentoringApplicationItem({
       </StyledApplicationInfoWrapper>
       <PhoneNumber status={status} phoneNumber={phoneNumber} />
       <StyledApplicationContent>{content}</StyledApplicationContent>
-      {status === '승인 대기' && (
-        <StyledButtonWrapper>
-          <StyledPrimaryButton>승인</StyledPrimaryButton>
-          <StyledSecondaryButton>거절</StyledSecondaryButton>
-        </StyledButtonWrapper>
-      )}
+      <ActionButtons status={status} />
     </StyledContainer>
   );
 }
@@ -88,30 +84,4 @@ const StyledApplicationFee = styled.p`
 const StyledApplicationContent = styled.p`
   color: ${({ theme }) => theme.FONT.B03};
   ${({ theme }) => theme.TYPOGRAPHY.B2_R}
-`;
-
-const StyledButtonWrapper = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 1rem;
-`;
-
-const StyledBaseButton = styled.button`
-  width: fit-content;
-  padding: 0.8rem 1.3rem;
-  border: none;
-  border-radius: 8px;
-
-  cursor: pointer;
-
-  color: ${({ theme }) => theme.FONT.W01};
-  ${({ theme }) => theme.TYPOGRAPHY.BTN4_R}
-`;
-
-const StyledPrimaryButton = styled(StyledBaseButton)`
-  background-color: ${({ theme }) => theme.SYSTEM.MAIN600};
-`;
-
-const StyledSecondaryButton = styled(StyledBaseButton)`
-  background-color: ${({ theme }) => theme.FONT.ERROR};
 `;
